refactor: separate chat commands in multiple files
