+++ src/view/chat/commands/disconnect.ts
@@ -1,0 +1,5 @@
+import { Socket } from 'socket.io'
+
+export function disconnectionCommand(socket: Socket): (message: any) => void {
+  return (message) => socket.disconnect()
+}
